Implementando botao calcular

+++ dstbin.cpp
@@ -53,7 +53,7 @@ void dstBin::calc()
         {
             int k = x;
             if (x > n-x) k = n-x;
-            unsigned long res = 1;
+            unsigned long long res = 1;
             for (int i = 1; i <= k; i++) res = (res*(n-k + i))/i;
             P[x] = res*pow(p, x)*pow(q, n-x);
         }
